+++ component/backend/Helper/Format.php
@@ -7,10 +7,12 @@
 
 namespace Akeeba\Subscriptions\Admin\Helper;
 
+use DateTimeZone;
 use FOF30\Container\Container;
 use FOF30\Date\Date;
 use FOF30\Model\DataModel;
 use JLoader;
+use Joomla\CMS\Factory;
 use JText;
 
 defined('_JEXEC') or die;
@@ -21,22 +23,48 @@ defined('_JEXEC') or die;
 abstract class Format
 {
 	/**
-	 * Format a date for display
+	 * Format a date for display.
 	 *
-	 * @param   string  $date    The date to format
-	 * @param   string  $format  The format string, default is whatever you specified in the component options
+	 * The $tzAware parameter defines whether the formatted date will be timezone-aware. If set to false the formatted
+	 * date will be rendered in the UTC timezone. If set to true the code will automatically try to use the logged in
+	 * user's timezone or, if none is set, the site's default timezone (Server Timezone). If set to a positive integer
+	 * the same thing will happen but for the specified user ID instead of the currently logged in user.
+	 *
+	 * @param   string    $date     The date to format
+	 * @param   string    $format   The format string, default is whatever you specified in the component options
+	 * @param   bool|int  $tzAware  Should the format be timezone aware? See notes above.
 	 *
 	 * @return string
 	 */
-	public static function date($date, $format = null)
+	public static function date($date, $format = null, $tzAware = true)
 	{
 		JLoader::import('joomla.utilities.date');
 
-		$jDate = new Date($date);
+		// Which timezone should I use?
+		$tz = null;
+
+		if ($tzAware !== false)
+		{
+			$userId    = is_bool($tzAware) ? null : (int) $tzAware;
+
+			try
+			{
+				$tzDefault = Factory::getApplication()->get('offset');
+			}
+			catch (\Exception $e)
+			{
+				$tzDefault = new DateTimeZone('GMT');
+			}
+
+			$user      = Factory::getUser($userId);
+			$tz        = $user->getParam('timezone', $tzDefault);
+		}
+
+		$jDate = new Date($date, $tz);
 
 		if (empty($format))
 		{
-			$format = self::getContainer()->params->get('dateformat', 'Y-m-d H:i');
+			$format = self::getContainer()->params->get('dateformat', 'Y-m-d H:i T');
 			$format = str_replace('%', '', $format);
 		}